add mocks, separate constants
