+++ src/app/core/constants.ts
@@ -1,1 +1,14 @@
 export const HUNDRED_TB = 109951162777600;
+export const HTTP_CODES = {
+  MAX_SPACE_USED: 420,
+  FORBIDDEN: 403,
+  NOT_FOUND: 404,
+};
+export enum ErrorMessages {
+  ServerUnavailable = 'Server Unavailable',
+  ServerError = 'Server Error',
+  InternalServerError = 'Internal Server Error',
+  NetworkError = 'Network Error',
+  ConnectionLost = 'Connection lost',
+  FilePickerCancelled = 'File picker was canceled or failed',
+}
