+++ src/app/core/services/http.service.ts
@@ -3,12 +3,6 @@ import axios, { AxiosRequestConfig, AxiosRequestHeaders } from 'axios';
 import packageJson from '../../../../package.json';
 import localStorageService from './local-storage.service';
 
-export const HTTP_CODES = {
-  MAX_SPACE_USED: 420,
-  FORBIDDEN: 403,
-  NOT_FOUND: 404,
-};
-
 const httpService = {
   get: <ResponseType>(url: string, config?: AxiosRequestConfig): Promise<ResponseType> => {
     return axios.get<ResponseType>(url, config).then((response) => response.data);
