+++ src/app/drive/services/downloadManager.service.ts
@@ -23,6 +23,7 @@ import { Iterator } from 'app/core/collections';
 import { SharedFiles, SharedFolders } from '@internxt/sdk/dist/drive/share/types';
 import { WorkspaceCredentialsDetails, WorkspaceData } from '@internxt/sdk/dist/workspaces';
 import { ConnectionLostError } from './../../network/requests';
+import { ErrorMessages } from 'app/core/constants';
 
 export type DownloadCredentials = {
   credentials: NetworkCredentials;
@@ -75,15 +76,6 @@ export type FileIterator = (
 
 export type SharedFolderIterator = (directoryId: string, resourcesToken?: string) => Iterator<SharedFolders>;
 export type SharedFileIterator = (directoryId: string, resourcesToken?: string) => Iterator<SharedFiles>;
-
-export enum ErrorMessages {
-  ServerUnavailable = 'Server Unavailable',
-  ServerError = 'Server Error',
-  InternalServerError = 'Internal Server Error',
-  NetworkError = 'Network Error',
-  ConnectionLost = 'Connection lost',
-  FilePickerCancelled = 'File picker was canceled or failed',
-}
 
 /**
  * DownloadManagerService handles file and folder downloads with queue management
