+++ src/app/share/views/ShareView/ShareFolderView.tsx
@@ -15,7 +15,7 @@ import { useEffect, useState } from 'react';
 import { match } from 'react-router';
 import { Link } from 'react-router-dom';
 import { WritableStream } from 'streamsaver';
-import { HTTP_CODES } from '../../../core/services/http.service';
+import { HTTP_CODES } from '../../../core/constants';
 import AppError from '../../../core/types';
 import { useAppSelector } from '../../../store/hooks';
 import SendBanner from './SendBanner';
